+++ packages/website/src/pages/index/subject/[id]/wiki/common.module.less
@@ -26,30 +26,9 @@
       height: 1.5rem;
     }
 
-    &Btn {
-      border: 2px dashed @gray-10;
-      border-radius: 12px;
-      width: 100%;
-      height: 44px;
-      font-size: 1rem;
-      color: @gray-60;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      gap: 10px;
-
-      &Secondary {
-        background: #fcfcfc;
-      }
-    }
-
     &Group {
       display: flex;
       flex: 1;
-
-      &Secondary * {
-        background: #fcfcfc;
-      }
     }
   }
 
@@ -80,7 +59,7 @@
     width: 120px;
   }
 
-  .Tips {
+  .tips {
     color: @gray-60;
     width: 100%;
     font-size: 0.75rem;
@@ -94,34 +73,6 @@
     flex: 1;
     gap: 10px;
     flex-direction: column;
-
-    &Item {
-      position: relative;
-      background-color: #fff;
-
-      &Minus {
-        display: none;
-        position: absolute;
-        right: 0;
-        top: 0;
-        cursor: pointer;
-        transform: translate(50%, -50%);
-      }
-
-      &:hover &Minus,
-      &:hover &Cursor {
-        display: block;
-      }
-
-      &Cursor {
-        display: none;
-        position: absolute;
-        left: 0;
-        top: 0;
-        cursor: move;
-        transform: translateY(20%);
-      }
-    }
   }
 }
 
@@ -162,8 +113,4 @@
   &More {
     align-self: end;
   }
-}
-
-.draggableBox {
-  margin-bottom: 0.625rem;
 }
